- moved code from AppLayoutRouterLayout to AppLayoutRouterLayoutBase and changed AppLayoutRouterLayout to be an intermediate class only defining a viewport to keep backwards compatibility
https://github.com/appreciated/vaadin-app-layout/issues/220
